requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
trino==0.328.0
redis==5.0.1
kafka-python==2.0.2
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pymysql==1.1.0
sshtunnel==0.4.0
cryptography==41.0.7
paramiko<3.0
pydantic[email]
aioredis
redis>=4.2.0rc1
passlib
bcrypt==4.0.1